Move Gradle runner JUnit extension to internal testing project
