+++ subprojects/build-adapter-xcode/src/functionalTest/java/dev/nokee/buildadapter/xcode/ConfigurationCacheDetectsXcodeProjectInitializationFunctionalTest.java
@@ -17,8 +17,8 @@ package dev.nokee.buildadapter.xcode;
 
 import dev.gradleplugins.runnerkit.BuildResult;
 import dev.gradleplugins.runnerkit.GradleRunner;
+import dev.nokee.internal.testing.junit.jupiter.ContextualGradleRunnerParameterResolver;
 import dev.nokee.platform.xcode.EmptyXCProject;
-import dev.nokee.platform.xcode.EmptyXCWorkspace;
 import net.nokeedev.testing.junit.jupiter.io.TestDirectory;
 import net.nokeedev.testing.junit.jupiter.io.TestDirectoryExtension;
 import org.junit.jupiter.api.BeforeAll;
