+++ subprojects/build-adapter-xcode/src/functionalTest/java/dev/nokee/buildadapter/xcode/XcodeWorkspaceToGradleProjectMappingFunctionalTest.java
@@ -16,6 +16,7 @@
 package dev.nokee.buildadapter.xcode;
 
 import dev.gradleplugins.runnerkit.GradleRunner;
+import dev.nokee.internal.testing.junit.jupiter.ContextualGradleRunnerParameterResolver;
 import dev.nokee.platform.xcode.EmptyXCProject;
 import dev.nokee.platform.xcode.XCWorkspaceElement;
 import net.nokeedev.testing.junit.jupiter.io.TestDirectory;
