+++ subprojects/internal-testing/src/main/java/dev/nokee/internal/testing/junit/jupiter/ContextualGradleRunnerParameterResolver.java
@@ -13,9 +13,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package dev.nokee.buildadapter.xcode;
+package dev.nokee.internal.testing.junit.jupiter;
 
-import dev.gradleplugins.runnerkit.GradleExecutor;
 import dev.gradleplugins.runnerkit.GradleRunner;
 import org.junit.jupiter.api.extension.ExtensionContext;
 import org.junit.jupiter.api.extension.ParameterContext;
